[DEVELOP][APK_GENERATOR] Add verbose for error when you generate apk/ add jinja to Makefile for more automatisation

+++ src/static/generate_apk/__init__.py
@@ -3,7 +3,7 @@ import shutil
 import subprocess
 from jinja2 import Template
 from pathlib import Path
-from log import warning
+from log import warning, error
 
 class GenerateAPK:
 
@@ -13,9 +13,20 @@ class GenerateAPK:
         if os.path.exists(self.__path):
             shutil.rmtree(self.__path)
         Path(self.__path).mkdir(parents=True, exist_ok=True)
-        shutil.copy(self.__base + "Makefile", self.__path + "Makefile")
+        
+        if args.sdktools and args.version_android:
+            makefile = JavaFile("Makefile",
+                                self.__base + "Makefile",
+                                {'version' : args.version_android,
+                                 'sdktools': args.sdktools})
+            with open(self.__path + "Makefile", 'w') as f:
+                f.write(makefile.get_content())
+        else:
+            shutil.copy(self.__base + "Makefile", self.__path + "Makefile")
+
+
         for javafile in javafiles:
-            file = self.__path + "/java/" +javafile.get_path()
+            file = self.__path + "/java/" + javafile.get_path()
             Path(os.path.dirname(file)).mkdir(parents=True, exist_ok=True)
             with open(file, 'w') as f:
                 f.write(javafile.get_content())
@@ -23,11 +34,32 @@ class GenerateAPK:
             f.write(manifest.get_content())
 
         if args.sdktools and args.version_android:
+            if not os.path.exists(f"{args.sdktools}"):
+                error("[APK_GENERATOR] The path specify for sdktools doesn't exist")
+                return
+            if not os.path.exists(f"{args.sdktools}/build-tools/"
+                                  f"{args.version_android}"):
+                error(f"[APK_GENERATOR] The path {args.sdktools}/build-tools/"
+                      f"{args.version_android} doesn't exist\n"
+                       "you should install it:\n"
+                       "\ttools/bin/sdkmanager --install "
+                      f"'build-tools;{args.version_android}'")
+                return
+            version_b = args.version_android.split(".")[0]
+            if not os.path.exists(f"{args.sdktools}/platforms/android-{version_b}"
+                                   "/android.jar"):
+                error(f"[APK_GENERATOR] The path {args.sdktools}/platforms/"
+                      f"android-{version_b} doesn't exist\n"
+                       "you should install it:\n"
+                       "\ttools/bin/sdkmanager --install "
+                      f"'platforms;android-{version_b}'")
+                return
+
             with open(self.__path + "/log.out", "w") as out:
                 with open(self.__path + "/log.err", "w") as err:
                     environ = os.environ
-                    environ["SDK_TOOLS"] = args.sdktools
-                    environ["VERSION"] = args.version_android
+                    #environ["SDK_TOOLS"] = args.sdktools
+                    #environ["VERSION"] = args.version_android
                     proc = subprocess.Popen(['make', '-C', self.__path],
                             env=environ,
                             stdout=out,
